backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic-settings>=2.0.0
httpx>=0.25.0
